Cart bộ lọc

+++ src/main/java/anbd/he191271/repository/VariantRepository.java
@@ -7,7 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 public interface VariantRepository extends JpaRepository<Variant, Integer> {
-    List<Variant> findByProductId(int productId);
+    List<Variant> findByProductId(long productId);
 //    @Transactional
 //    void deleteByProductId(int productId);
 }
